+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Notable changes only.
 
-## [unreleased]
+## [0.5.1] - 2024-08-02
 
 ### Added
 
@@ -110,7 +110,8 @@ Most of those addition are breaking because they shadows `str`'s methods.
 
 Initial release
 
-[unreleased]: https://github.com/polazarus/hipstr/compare/0.4.0...HEAD
+<!-- [unreleased]: https://github.com/polazarus/hipstr/compare/0.5.1...HEAD -->
+[0.5.1]: https://github.com/polazarus/hipstr/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/polazarus/hipstr/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/polazarus/hipstr/compare/0.3.3...0.4.0
 [0.3.3]: https://github.com/polazarus/hipstr/compare/0.3.2...0.3.3